feat(nimbus): Support Firefox Labs on Fenix
Because:

- Fenix supports Firefox Labs as of Firefox 154+;
- the constants the controlled in what applications supported Firefox
  Labs were spread across `NimbusConstants` instead of being associated
  with the application configuuration; and
- Firefox version constants were declared quite late in
  `NimbusConstants`, making them unavailable to `ApplicationConfig`
  inststances

this commit:

- refactors `NimbusConstants.Version` into
  `experimenter.experiments.versions`;
- refactors all the Firefox Labs fields in `NimbusConstants` into a new
  dataclass that is used with `ApplicationConfig`;
- updates Firefox Desktop and Fenix's application configs to use the new
  FirefoxLabs dataclass;
- updates branches form rendering to conditionally render Firefox Labs
  fields when supported in the application config;
- updates available choices for firefox_labs_group to be based on the
  application; and
- updates the existing unit tests to be generic across all Firefox
  Labs-supporting applications.

Fixes #16199
