+++ experimenter/experimenter/experiments/versions.py
@@ -1,0 +1,389 @@
+import packaging.version
+from django.db import models
+
+
+class Version(models.TextChoices):
+    @staticmethod
+    def parse(version_str: str) -> packaging.version.Version:
+        return packaging.version.parse(version_str.replace("!", "0"))
+
+    NO_VERSION = ""
+    FIREFOX_11 = "11.!"
+    FIREFOX_12 = "12.!"
+    FIREFOX_13 = "13.!"
+    FIREFOX_14 = "14.!"
+    FIREFOX_15 = "15.!"
+    FIREFOX_16 = "16.!"
+    FIREFOX_17 = "17.!"
+    FIREFOX_18 = "18.!"
+    FIREFOX_19 = "19.!"
+    FIREFOX_20 = "20.!"
+    FIREFOX_21 = "21.!"
+    FIREFOX_22 = "22.!"
+    FIREFOX_23 = "23.!"
+    FIREFOX_24 = "24.!"
+    FIREFOX_25 = "25.!"
+    FIREFOX_26 = "26.!"
+    FIREFOX_27 = "27.!"
+    FIREFOX_28 = "28.!"
+    FIREFOX_29 = "29.!"
+    FIREFOX_30 = "30.!"
+    FIREFOX_31 = "31.!"
+    FIREFOX_32 = "32.!"
+    FIREFOX_33 = "33.!"
+    FIREFOX_34 = "34.!"
+    FIREFOX_35 = "35.!"
+    FIREFOX_36 = "36.!"
+    FIREFOX_37 = "37.!"
+    FIREFOX_38 = "38.!"
+    FIREFOX_39 = "39.!"
+    FIREFOX_40 = "40.!"
+    FIREFOX_41 = "41.!"
+    FIREFOX_42 = "42.!"
+    FIREFOX_43 = "43.!"
+    FIREFOX_44 = "44.!"
+    FIREFOX_45 = "45.!"
+    FIREFOX_46 = "46.!"
+    FIREFOX_47 = "47.!"
+    FIREFOX_48 = "48.!"
+    FIREFOX_49 = "49.!"
+    FIREFOX_50 = "50.!"
+    FIREFOX_51 = "51.!"
+    FIREFOX_52 = "52.!"
+    FIREFOX_53 = "53.!"
+    FIREFOX_54 = "54.!"
+    FIREFOX_55 = "55.!"
+    FIREFOX_56 = "56.!"
+    FIREFOX_57 = "57.!"
+    FIREFOX_58 = "58.!"
+    FIREFOX_59 = "59.!"
+    FIREFOX_60 = "60.!"
+    FIREFOX_61 = "61.!"
+    FIREFOX_62 = "62.!"
+    FIREFOX_63 = "63.!"
+    FIREFOX_64 = "64.!"
+    FIREFOX_65 = "65.!"
+    FIREFOX_66 = "66.!"
+    FIREFOX_67 = "67.!"
+    FIREFOX_68 = "68.!"
+    FIREFOX_69 = "69.!"
+    FIREFOX_70 = "70.!"
+    FIREFOX_71 = "71.!"
+    FIREFOX_72 = "72.!"
+    FIREFOX_73 = "73.!"
+    FIREFOX_74 = "74.!"
+    FIREFOX_75 = "75.!"
+    FIREFOX_76 = "76.!"
+    FIREFOX_77 = "77.!"
+    FIREFOX_78 = "78.!"
+    FIREFOX_79 = "79.!"
+    FIREFOX_80 = "80.!"
+    FIREFOX_81 = "81.!"
+    FIREFOX_82 = "82.!"
+    FIREFOX_83 = "83.!"
+    FIREFOX_84 = "84.!"
+    FIREFOX_85 = "85.!"
+    FIREFOX_86 = "86.!"
+    FIREFOX_87 = "87.!"
+    FIREFOX_88 = "88.!"
+    FIREFOX_89 = "89.!"
+    FIREFOX_90 = "90.!"
+    FIREFOX_91 = "91.!"
+    FIREFOX_92 = "92.!"
+    FIREFOX_9201 = "92.0.1"
+    FIREFOX_93 = "93.!"
+    FIREFOX_94 = "94.!"
+    FIREFOX_95 = "95.!"
+    FIREFOX_96 = "96.!"
+    FIREFOX_9601 = "96.0.1"
+    FIREFOX_9602 = "96.0.2"
+    FIREFOX_97 = "97.!"
+    FIREFOX_98 = "98.!"
+    FIREFOX_9830 = "98.3.0"
+    FIREFOX_99 = "99.!"
+    FIREFOX_9910 = "99.1.0"
+    FIREFOX_100 = "100.!"
+    FIREFOX_101 = "101.!"
+    FIREFOX_102 = "102.!"
+    FIREFOX_103 = "103.!"
+    FIREFOX_10301 = "103.0.1"
+    FIREFOX_104 = "104.!"
+    FIREFOX_105 = "105.!"
+    FIREFOX_10501 = "105.0.1"
+    FIREFOX_10502 = "105.0.2"
+    FIREFOX_10503 = "105.0.3"
+    FIREFOX_106 = "106.!"
+    FIREFOX_10601 = "106.0.1"
+    FIREFOX_10602 = "106.0.2"
+    FIREFOX_107 = "107.!"
+    FIREFOX_108 = "108.!"
+    FIREFOX_109 = "109.!"
+    FIREFOX_110 = "110.!"
+    FIREFOX_111 = "111.!"
+    FIREFOX_111_0_1 = "111.0.1"
+    FIREFOX_112 = "112.!"
+    FIREFOX_113 = "113.!"
+    FIREFOX_113_0_1 = "113.0.1"
+    FIREFOX_114 = "114.!"
+    FIREFOX_114_3_0 = "114.3.0"
+    FIREFOX_115 = "115.!"
+    FIREFOX_115_0_2 = "115.0.2"
+    FIREFOX_115_7 = "115.7.0"
+    FIREFOX_115_25 = "115.25.0"
+    FIREFOX_116 = "116.!"
+    FIREFOX_116_0_1 = "116.0.1"
+    FIREFOX_116_2_0 = "116.2.0"
+    FIREFOX_116_3_0 = "116.3.0"
+    FIREFOX_117 = "117.!"
+    FIREFOX_118 = "118.!"
+    FIREFOX_118_0_1 = "118.0.1"
+    FIREFOX_118_0_2 = "118.0.2"
+    FIREFOX_119 = "119.!"
+    FIREFOX_120 = "120.!"
+    FIREFOX_121 = "121.!"
+    FIREFOX_121_0_1 = "121.0.1"
+    FIREFOX_122 = "122.!"
+    FIREFOX_122_1_0 = "122.1.0"
+    FIREFOX_122_2_0 = "122.2.0"
+    FIREFOX_123 = "123.!"
+    FIREFOX_123_0_1 = "123.0.1"
+    FIREFOX_124 = "124.!"
+    FIREFOX_124_2_0 = "124.2.0"
+    FIREFOX_124_3_0 = "124.3.0"
+    FIREFOX_125 = "125.!"
+    FIREFOX_125_0_1 = "125.0.1"
+    FIREFOX_125_0_2 = "125.0.2"
+    FIREFOX_125_1_0 = "125.1.0"
+    FIREFOX_125_2_0 = "125.2.0"
+    FIREFOX_126 = "126.!"
+    FIREFOX_126_1_0 = "126.1.0"
+    FIREFOX_126_2_0 = "126.2.0"
+    FIREFOX_127 = "127.!"
+    FIREFOX_127_0_1 = "127.0.1"
+    FIREFOX_127_0_2 = "127.0.2"
+    FIREFOX_128 = "128.!"
+    FIREFOX_128_12 = "128.12.0"
+    FIREFOX_129 = "129.!"
+    FIREFOX_130 = "130.!"
+    FIREFOX_130_0_1 = "130.0.1"
+    FIREFOX_131 = "131.!"
+    FIREFOX_131_B4 = "131.0b4"
+    FIREFOX_131_0_3 = "131.0.3"
+    FIREFOX_131_1_0 = "131.1.0"
+    FIREFOX_131_2_0 = "131.2.0"
+    FIREFOX_132 = "132.!"
+    FIREFOX_132_B6 = "132.0b6"
+    FIREFOX_133 = "133.!"
+    FIREFOX_133_B8 = "133.0b8"
+    FIREFOX_133_0_1 = "133.0.1"
+    FIREFOX_134 = "134.!"
+    FIREFOX_134_1_0 = "134.1.0"
+    FIREFOX_135 = "135.!"
+    FIREFOX_135_0_1 = "135.0.1"
+    FIREFOX_135_1_0 = "135.1.0"
+    FIREFOX_136 = "136.!"
+    FIREFOX_136_0_2 = "136.0.2"
+    FIREFOX_137 = "137.!"
+    FIREFOX_137_1_0 = "137.1.0"
+    FIREFOX_137_2_0 = "137.2.0"
+    FIREFOX_138 = "138.!"
+    FIREFOX_138_B3 = "138.0b3"
+    FIREFOX_138_1_0 = "138.1.0"
+    FIREFOX_138_2_0 = "138.2.0"
+    FIREFOX_138_0_3 = "138.0.3"
+    FIREFOX_139 = "139.!"
+    FIREFOX_139_1_0 = "139.1.0"
+    FIREFOX_139_2_0 = "139.2.0"
+    FIREFOX_139_0_4 = "139.0.4"
+    FIREFOX_140 = "140.!"
+    FIREFOX_140_0_1 = "140.0.1"
+    FIREFOX_140_0_2 = "140.0.2"
+    FIREFOX_140_0_3 = "140.0.3"
+    FIREFOX_140_0_4 = "140.0.4"
+    FIREFOX_140_1_0 = "140.1.0"
+    FIREFOX_140_2_0 = "140.2.0"
+    FIREFOX_140_3_0 = "140.3.0"
+    FIREFOX_140_4_0 = "140.4.0"
+    FIREFOX_141 = "141.!"
+    FIREFOX_141_0_1 = "141.0.1"
+    FIREFOX_141_0_2 = "141.0.2"
+    FIREFOX_141_0_3 = "141.0.3"
+    FIREFOX_141_0_4 = "141.0.4"
+    FIREFOX_141_1_0 = "141.1.0"
+    FIREFOX_141_2_0 = "141.2.0"
+    FIREFOX_141_3_0 = "141.3.0"
+    FIREFOX_141_4_0 = "141.4.0"
+    FIREFOX_142 = "142.!"
+    FIREFOX_142_0_1 = "142.0.1"
+    FIREFOX_142_0_2 = "142.0.2"
+    FIREFOX_142_0_3 = "142.0.3"
+    FIREFOX_142_0_4 = "142.0.4"
+    FIREFOX_142_1_0 = "142.1.0"
+    FIREFOX_142_2_0 = "142.2.0"
+    FIREFOX_142_3_0 = "142.3.0"
+    FIREFOX_142_4_0 = "142.4.0"
+    FIREFOX_143 = "143.!"
+    FIREFOX_143_B3 = "143.0b3"
+    FIREFOX_143_B4 = "143.0b4"
+    FIREFOX_143_0_1 = "143.0.1"
+    FIREFOX_143_0_2 = "143.0.2"
+    FIREFOX_143_0_3 = "143.0.3"
+    FIREFOX_143_0_4 = "143.0.4"
+    FIREFOX_143_1_0 = "143.1.0"
+    FIREFOX_143_1_1 = "143.1.1"
+    FIREFOX_143_2_0 = "143.2.0"
+    FIREFOX_143_3_0 = "143.3.0"
+    FIREFOX_143_4_0 = "143.4.0"
+    FIREFOX_144 = "144.!"
+    FIREFOX_144_0_1 = "144.0.1"
+    FIREFOX_144_0_2 = "144.0.2"
+    FIREFOX_144_0_3 = "144.0.3"
+    FIREFOX_144_0_4 = "144.0.4"
+    FIREFOX_144_1_0 = "144.1.0"
+    FIREFOX_144_2_0 = "144.2.0"
+    FIREFOX_144_3_0 = "144.3.0"
+    FIREFOX_144_4_0 = "144.4.0"
+    FIREFOX_145 = "145.!"
+    FIREFOX_145_0_1 = "145.0.1"
+    FIREFOX_145_0_2 = "145.0.2"
+    FIREFOX_145_0_3 = "145.0.3"
+    FIREFOX_145_0_4 = "145.0.4"
+    FIREFOX_145_1_0 = "145.1.0"
+    FIREFOX_145_2_0 = "145.2.0"
+    FIREFOX_145_3_0 = "145.3.0"
+    FIREFOX_145_4_0 = "145.4.0"
+    FIREFOX_146 = "146.!"
+    FIREFOX_146_0_1 = "146.0.1"
+    FIREFOX_146_0_2 = "146.0.2"
+    FIREFOX_146_0_3 = "146.0.3"
+    FIREFOX_146_0_4 = "146.0.4"
+    FIREFOX_146_1_0 = "146.1.0"
+    FIREFOX_146_2_0 = "146.2.0"
+    FIREFOX_146_3_0 = "146.3.0"
+    FIREFOX_146_4_0 = "146.4.0"
+    FIREFOX_147 = "147.!"
+    FIREFOX_147_0_1 = "147.0.1"
+    FIREFOX_147_0_2 = "147.0.2"
+    FIREFOX_147_0_3 = "147.0.3"
+    FIREFOX_147_0_4 = "147.0.4"
+    FIREFOX_147_1_0 = "147.1.0"
+    FIREFOX_147_2_0 = "147.2.0"
+    FIREFOX_147_3_0 = "147.3.0"
+    FIREFOX_147_4_0 = "147.4.0"
+    FIREFOX_148 = "148.!"
+    FIREFOX_148_0_1 = "148.0.1"
+    FIREFOX_148_0_2 = "148.0.2"
+    FIREFOX_148_0_3 = "148.0.3"
+    FIREFOX_148_0_4 = "148.0.4"
+    FIREFOX_148_1_0 = "148.1.0"
+    FIREFOX_148_2_0 = "148.2.0"
+    FIREFOX_148_3_0 = "148.3.0"
+    FIREFOX_148_4_0 = "148.4.0"
+    FIREFOX_149 = "149.!"
+    FIREFOX_149_0_1 = "149.0.1"
+    FIREFOX_149_0_2 = "149.0.2"
+    FIREFOX_149_0_3 = "149.0.3"
+    FIREFOX_149_0_4 = "149.0.4"
+    FIREFOX_149_1_0 = "149.1.0"
+    FIREFOX_149_2_0 = "149.2.0"
+    FIREFOX_149_3_0 = "149.3.0"
+    FIREFOX_149_4_0 = "149.4.0"
+    FIREFOX_150 = "150.!"
+    FIREFOX_150_0_1 = "150.0.1"
+    FIREFOX_150_0_2 = "150.0.2"
+    FIREFOX_150_0_3 = "150.0.3"
+    FIREFOX_150_0_4 = "150.0.4"
+    FIREFOX_150_1_0 = "150.1.0"
+    FIREFOX_150_2_0 = "150.2.0"
+    FIREFOX_150_3_0 = "150.3.0"
+    FIREFOX_150_4_0 = "150.4.0"
+    FIREFOX_151 = "151.!"
+    FIREFOX_151_0_1 = "151.0.1"
+    FIREFOX_151_0_2 = "151.0.2"
+    FIREFOX_151_0_3 = "151.0.3"
+    FIREFOX_151_0_4 = "151.0.4"
+    FIREFOX_151_1_0 = "151.1.0"
+    FIREFOX_151_2_0 = "151.2.0"
+    FIREFOX_151_3_0 = "151.3.0"
+    FIREFOX_151_4_0 = "151.4.0"
+    FIREFOX_152 = "152.!"
+    FIREFOX_152_0_1 = "152.0.1"
+    FIREFOX_152_0_2 = "152.0.2"
+    FIREFOX_152_0_3 = "152.0.3"
+    FIREFOX_152_0_4 = "152.0.4"
+    FIREFOX_152_1_0 = "152.1.0"
+    FIREFOX_152_2_0 = "152.2.0"
+    FIREFOX_152_3_0 = "152.3.0"
+    FIREFOX_152_4_0 = "152.4.0"
+    FIREFOX_153 = "153.!"
+    FIREFOX_153_0_1 = "153.0.1"
+    FIREFOX_153_0_2 = "153.0.2"
+    FIREFOX_153_0_3 = "153.0.3"
+    FIREFOX_153_0_4 = "153.0.4"
+    FIREFOX_153_1_0 = "153.1.0"
+    FIREFOX_153_2_0 = "153.2.0"
+    FIREFOX_153_3_0 = "153.3.0"
+    FIREFOX_153_4_0 = "153.4.0"
+    FIREFOX_154 = "154.!"
+    FIREFOX_154_0_1 = "154.0.1"
+    FIREFOX_154_0_2 = "154.0.2"
+    FIREFOX_154_0_3 = "154.0.3"
+    FIREFOX_154_0_4 = "154.0.4"
+    FIREFOX_154_1_0 = "154.1.0"
+    FIREFOX_154_2_0 = "154.2.0"
+    FIREFOX_154_3_0 = "154.3.0"
+    FIREFOX_154_4_0 = "154.4.0"
+    FIREFOX_155 = "155.!"
+    FIREFOX_155_0_1 = "155.0.1"
+    FIREFOX_155_0_2 = "155.0.2"
+    FIREFOX_155_0_3 = "155.0.3"
+    FIREFOX_155_0_4 = "155.0.4"
+    FIREFOX_155_1_0 = "155.1.0"
+    FIREFOX_155_2_0 = "155.2.0"
+    FIREFOX_155_3_0 = "155.3.0"
+    FIREFOX_155_4_0 = "155.4.0"
+    FIREFOX_156 = "156.!"
+    FIREFOX_156_0_1 = "156.0.1"
+    FIREFOX_156_0_2 = "156.0.2"
+    FIREFOX_156_0_3 = "156.0.3"
+    FIREFOX_156_0_4 = "156.0.4"
+    FIREFOX_156_1_0 = "156.1.0"
+    FIREFOX_156_2_0 = "156.2.0"
+    FIREFOX_156_3_0 = "156.3.0"
+    FIREFOX_156_4_0 = "156.4.0"
+    FIREFOX_157 = "157.!"
+    FIREFOX_157_0_1 = "157.0.1"
+    FIREFOX_157_0_2 = "157.0.2"
+    FIREFOX_157_0_3 = "157.0.3"
+    FIREFOX_157_0_4 = "157.0.4"
+    FIREFOX_157_1_0 = "157.1.0"
+    FIREFOX_157_2_0 = "157.2.0"
+    FIREFOX_157_3_0 = "157.3.0"
+    FIREFOX_157_4_0 = "157.4.0"
+    FIREFOX_158 = "158.!"
+    FIREFOX_158_0_1 = "158.0.1"
+    FIREFOX_158_0_2 = "158.0.2"
+    FIREFOX_158_0_3 = "158.0.3"
+    FIREFOX_158_0_4 = "158.0.4"
+    FIREFOX_158_1_0 = "158.1.0"
+    FIREFOX_158_2_0 = "158.2.0"
+    FIREFOX_158_3_0 = "158.3.0"
+    FIREFOX_158_4_0 = "158.4.0"
+    FIREFOX_159 = "159.!"
+    FIREFOX_159_0_1 = "159.0.1"
+    FIREFOX_159_0_2 = "159.0.2"
+    FIREFOX_159_0_3 = "159.0.3"
+    FIREFOX_159_0_4 = "159.0.4"
+    FIREFOX_159_1_0 = "159.1.0"
+    FIREFOX_159_2_0 = "159.2.0"
+    FIREFOX_159_3_0 = "159.3.0"
+    FIREFOX_159_4_0 = "159.4.0"
+    FIREFOX_160 = "160.!"
+    FIREFOX_160_0_1 = "160.0.1"
+    FIREFOX_160_0_2 = "160.0.2"
+    FIREFOX_160_0_3 = "160.0.3"
+    FIREFOX_160_0_4 = "160.0.4"
+    FIREFOX_160_1_0 = "160.1.0"
+    FIREFOX_160_2_0 = "160.2.0"
+    FIREFOX_160_3_0 = "160.3.0"
+    FIREFOX_160_4_0 = "160.4.0"
