+++ experimenter/pyproject.toml
@@ -77,6 +77,7 @@ markdown = "^3.10.2"
 nh3 = "^0.3"
 django-prose = "^2.1.0"
 scipy = "^1.15.0"
+typing-extensions = "^4.16.0"
 
 [build-system]
 requires = ["poetry>=2.1.1"]
